+++ src/main/java/net/geforcemods/securitycraft/misc/CustomDamageSources.java
@@ -1,6 +1,8 @@
 package net.geforcemods.securitycraft.misc;
 
+import net.minecraft.entity.Entity;
 import net.minecraft.util.DamageSource;
+import net.minecraft.util.EntityDamageSource;
 
 /**
  * SecurityCraft's custom damage source handler.
@@ -11,6 +13,20 @@ public class CustomDamageSources {
 	public static final DamageSource LASER = new DamageSource("securitycraft.laser");
 	public static final DamageSource FAKE_WATER = new DamageSource("securitycraft.fakeWater").setDamageBypassesArmor();
 	public static final DamageSource ELECTRICITY = new DamageSource("securitycraft.electricity").setDamageBypassesArmor();
-	public static final DamageSource TASER = new DamageSource("securitycraft.taser");
 	public static final DamageSource INCORRECT_PASSCODE = new DamageSource("securitycraft.incorrectPasscode");
+
+	public static DamageSource taser(Entity shooter) {
+		return new TaserDamageSource("securitycraft.taser", shooter);
+	}
+
+	static class TaserDamageSource extends EntityDamageSource {
+		public TaserDamageSource(String damageTypeId, Entity shooter) {
+			super(damageTypeId, shooter);
+		}
+
+		@Override
+		public Entity getImmediateSource() {
+			return null;
+		}
+	}
 }