+++ Changelog.md
@@ -1,6 +1,7 @@
 --------------------------Changelog for v1.9.7 of SecurityCraft--------------------------
 
 - Fix: Entities around a player sometimes get invisible when the player dismounts a camera
+- Fix: Shields do not take damage when blocking a taser that deals high enough damage
 - Fix: The Sentry-/Mine Remote Access Tool screens do not show the item's custom name
 
 --------------------------Changelog for v1.9.6 of SecurityCraft--------------------------